Add integrationtests for terminate end events
