+++ _integration_tests/config/test/iam/claim_service/claims/default_namespace/default_permissions.json
