+++ _integration_tests/config/test/iam/jwt_token_adapter.json
