+++ _integration_tests/config/test/messagebus/faye.json
