+++ _integration_tests/test/terminate_end_event.js
@@ -1,31 +1,29 @@
 'use strict';
 
 const should = require('should');
-const setup = require('../application/test_setup');
+const ProcessEngineServiceTestFixture = require('../dist/commonjs/process_engine_service_test_fixture').ProcessEngineServiceTestFixture;
+
+const testTimeoutInMS = 5000;
 
 describe.only('Terminate End Event', function () {
 
-  let httpBootstrapper;
-  let dummyContext;
-  let processEngineService;
-  this.timeout(5000);
+  let processEngineServiceFixture;
+
+  this.timeout(testTimeoutInMS);
 
   before(async () => {
-    httpBootstrapper = await setup.initializeBootstrapper();
-    dummyContext = await setup.createContext();
-    processEngineService = await setup.resolveAsync('ProcessEngineService');
+    processEngineServiceFixture = new ProcessEngineServiceTestFixture();
+    await processEngineServiceFixture.setup();
   });
 
   after(async () => {
-    await httpBootstrapper.reset();
-    await httpBootstrapper.shutdown();
+    await processEngineServiceFixture.tearDown();
   });
 
   it(`should successfully terminate a process upon reaching a TerminateEndEvent.`, async () => {
 
     const processKey = 'terminate_end_event_sample';
-    const initialToken = {};
-    const result = await processEngineService.executeProcess(dummyContext, undefined, processKey, initialToken);
+    const result = await processEngineServiceFixture.executeProcess(processKey);
 
     const expectedResult = /terminated/i;
